Added BigPoint Class
The BigPoint class is a replica of the Java Point class, however it does not extend Point2D because the methods that must be implemented from that extension do not translate well to this class, due to narrowing conversions. It will be used in place of Point in order to ensure that cells being very far away from the origin do not cause crashes, however this may cause overhead that could possibly lead to poor performance down the line. Perhaps a setting could be made to switch between the two, with a boundary or toroidal behavior avaliable for the implementation using the Java Point class when cells approach the primitive integer limit.