BigPoint class removed
The idea for extending the plane beyond the integer limit hit a snag when I realized that the draw functions in the java.awt.Graphics class only accept integers. This means that unless there is an alternative way to draw, there is
no point in tracking the values of various things with BigIntegers.